fix: hilt config
