+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,6 @@
         <activity
             android:name=".ui.sheet.SheetActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.Sheet">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
