+++ app/src/main/java/com/paixao/labs/myapplication/App.kt
@@ -3,12 +3,12 @@ package com.paixao.labs.myapplication
 import android.app.Application
 import com.google.firebase.FirebaseApp
 import dagger.hilt.android.HiltAndroidApp
+import javax.inject.Inject
 
 @HiltAndroidApp
 class App : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        FirebaseApp.initializeApp(this.applicationContext)
     }
 }
