+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">sheet</string>
+    <string name="app_name">HotBattle RPG</string>
     <string name="sheet_title">Ficha de %s</string>
 </resources>
