+++ build.gradle
@@ -1,20 +1,30 @@
 buildscript {
+
+    repositories {
+        google()
+        mavenCentral()
+        gradlePluginPortal()
+        maven { url "https://maven.pkg.jetbrains.space/public/p/compose/dev" }
+    }
+
     ext {
         compose_version = '1.1.0-beta01'
+        composeCompiler = '1.1.0-rc02'
         hilt_version = '2.40.5'
+        voyager_version = '1.0.0-beta15'
     }
 
     dependencies {
         classpath 'com.google.gms:google-services:4.3.10'
         classpath 'com.google.dagger:hilt-android-gradle-plugin:2.40.5'
+        classpath 'org.jetbrains.compose:compose-gradle-plugin:1.0.1-rc2'
     }
 }
 plugins {
     id 'com.android.application' version '7.2.0-alpha06' apply false
     id 'com.android.library' version '7.2.0-alpha06' apply false
-    id 'org.jetbrains.kotlin.android' version '1.5.31' apply false
+    id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
     id 'org.jmailen.kotlinter' version "3.3.0" apply true
-
 }
 
 task clean(type: Delete) {
